+++ src/DotNext.Tests/IO/StreamSourceTests.cs
@@ -608,7 +608,7 @@ public sealed class StreamSourceTests : Test
     [Fact]
     public static async Task SharedStreamConcurrentReadAsync()
     {
-        byte[] expected = [10, 20, 30, 40, 50, 60];
+        byte[] expected = RandomBytes(512);
 
         await using var stream = StreamSource.AsSharedStream(new(expected));
 
@@ -632,12 +632,14 @@ public sealed class StreamSourceTests : Test
         }
     }
 
-    [Fact]
-    public static void SharedStreamConcurrentRead()
+    [Theory]
+    [InlineData(false)]
+    [InlineData(true)]
+    public static void SharedStreamConcurrentRead(bool compatWithAsync)
     {
-        byte[] expected = [10, 20, 30, 40, 50, 60];
+        byte[] expected = RandomBytes(512);
 
-        using var stream = StreamSource.AsSharedStream(new(expected));
+        using var stream = StreamSource.AsSharedStream(new(expected), compatWithAsync);
 
         var thread1 = new Thread(ReadStream);
         var thread2 = new Thread(ReadStream);